Accumulate mouse motion as float, SDL3 reports subpixel relative motion
Truncating SDL3's float xrel/yrel per event loses fractional motion with
a bias toward zero, which shows up as notchy low speed aim. Also remove
an unused SDL2 event shim and duplicated defines.

+++ Quake/input.h
@@ -32,7 +32,7 @@ void IN_Commands (void);
 // oportunity for devices to stick commands on the script buffer
 
 // mouse moved by dx and dy pixels
-void IN_MouseMotion (int dx, int dy);
+void IN_MouseMotion (float dx, float dy);
 
 void IN_SendKeyEvents (void);
 // used as a callback for Sys_SendKeyEvents() by some drivers
